@@ -3,7 +3,7 @@ class Keyword < ActiveRecord::Base
   
   belongs_to :meta_term, :class_name => "MetaTerm"
   belongs_to :user # TODO person ??
-  #belongs_to :media_entry
+  belongs_to :meta_datum
 
   validates_presence_of :meta_term
 
@@ -13,23 +13,6 @@ class Keyword < ActiveRecord::Base
     "#{meta_term}"
   end
   
-  #tmp# wrong! TODO through new method meta_data
-#  def resources
-#    a = []
-#    MetaKey.where(:object_type => "Keyword").each do |key|
-#      key.meta_data.each do |md|
-#        a += [md.resource_type, md.resource_id] if md.value.include?(id)
-#      end
-#    end
-#    a
-#  end
-
-  def meta_data
-    MetaDatum.joins(:meta_key).
-      where(:meta_keys => {:object_type => self.class.name}).
-      where(["value REGEXP ?", "-\ #{id}\n" ])
-  end
-
 #######################################
 
   def self.search(query)
