@@ -1,6 +1,6 @@
 # -*- encoding : utf-8 -*-
  
-class MetaDatumString < MetaDatumBase
+class MetaDatumString < MetaDatum
 
   alias_attribute :value, :string
 
@@ -8,14 +8,17 @@ class MetaDatumString < MetaDatumBase
     SQLHelper.execute_sql "UPDATE meta_data SET value = NULL where id = #{id}"
   end
 
+
   def set_value_before_save
   end
 
-  def deserialized_value
+  def deserialized_value(user = nil)
+    # TODO super if meta_key.is_dynamic? # when MetaDatum is the superclass
     string
   end
 
   def to_s
+    # TODO super # when MetaDatum is the superclass
     string
   end
 
