@@ -36,9 +36,9 @@ MAdeK::Application.configure do
   # like if you have constraints or database-specific column types
   config.active_record.schema_format = :sql
 
+  # Raise exception on mass assignment protection for Active Record models
+  config.active_record.mass_assignment_sanitizer = :strict
+
   # Print deprecation notices to the stderr
   config.active_support.deprecation = :stderr
-
-  # For JavaScript testing in the browser
-  config.use_transactional_fixtures = false
 end
