@@ -43,6 +43,7 @@ end
 When /^(?:|I )select "(.+)" as parent set$/ do |label|
   label.gsub!(/\s/, "_")
   raise "#{label} is already selected so you can not select it again" if find("input##{label}").checked?
+  wait_until(20) { find("input##{label}:not(selected)") }
   find("input##{label}:not(selected)").click
   raise "#{label} was not selected" unless find("input##{label}").checked?
 end
@@ -55,9 +56,14 @@ When /^(?:|I )deselect "(.+)" as parent set$/ do |label|
 end
 
 When /^(?:|I )submit the selection widget$/ do
-  wait_until(25){find(".widget .submit")}
+  wait_until(40){find(".widget .submit")}
   find(".widget .submit").click
-  wait_until(25){ all(".widget", :visible => true).size == 0 }
+  
+  # This always breaks on CI but works locally -- no idea how to really solve this.
+  # BOTH of these time out -- let's try a super-idiotic approach below
+  #wait_until(40){ all(".set.widget", :visible => true).size == 0 }
+  #wait_until(55){ all(".set.widget").size == 0 }
+  sleep 5.0 # Super-idiotic approach
 end
 
 When /^(?:|I )create a new set named "(.+)"$/ do |name|
