@@ -4,6 +4,7 @@ describe MediaResourcesController do
   render_views
   
   before :all do
+    DevelopmentHelpers::MetaDataPreset.load_minimal_yaml
     @user = FactoryGirl.create :user
   end
   
@@ -14,7 +15,8 @@ describe MediaResourcesController do
         type = rand > 0.5 ? :media_entry : :media_set
         mr = Factory type, :user => @user
         mr.parents << FactoryGirl.create(:media_set, :user => @user)
-        mr.meta_data.create(:meta_key => MetaKey.find_by_label("title"), :value => Faker::Lorem.words(4).join(' '))
+        mr.meta_data.create(:meta_key => MetaKey.find_by_label("title"), 
+                            :value => Faker::Lorem.words(4).join(' '))
       end
       # MetaContext
       @meta_context = MetaContext.first
