@@ -1,19 +1,19 @@
 require 'spec_helper'
 
-describe MetaDatumPerson do
+describe MetaDatumPeople do
 
   describe "Creation" do
 
     it "should not raise an error " do
-      expect {FactoryGirl.create :meta_datum_person}.not_to raise_error
+      expect {FactoryGirl.create :meta_datum_people}.not_to raise_error
     end
 
     it "should not be nil" do
-      (FactoryGirl.create :meta_datum_person).should_not == nil
+      (FactoryGirl.create :meta_datum_people).should_not == nil
     end
 
     it "should be persisted" do
-      (FactoryGirl.create :meta_datum_person).should be_persisted
+      (FactoryGirl.create :meta_datum_people).should be_persisted
     end
 
   end
@@ -22,7 +22,7 @@ describe MetaDatumPerson do
   describe "Linking with People" do
 
     before :each do
-      @mdp = FactoryGirl.create :meta_datum_person
+      @mdp = FactoryGirl.create :meta_datum_people
       @person1 = FactoryGirl.create :person
       @person2 = FactoryGirl.create :person
     end
@@ -34,16 +34,16 @@ describe MetaDatumPerson do
      it "should have persist added relations" do
       @mdp.people << @person1
       @mdp.people << @person2
-      MetaDatumPerson.find(@mdp.id).people.should include @person1
-      MetaDatumPerson.find(@mdp.id).people.should include @person2
+      MetaDatumPeople.find(@mdp.id).people.should include @person1
+      MetaDatumPeople.find(@mdp.id).people.should include @person2
     end
 
     describe "value interface" do
       it "should be an alias for people" do
         @mdp.people << @person1
         @mdp.people << @person2
-        MetaDatumPerson.find(@mdp.id).value.should include @person1
-        MetaDatumPerson.find(@mdp.id).value.should include @person2
+        MetaDatumPeople.find(@mdp.id).value.should include @person1
+        MetaDatumPeople.find(@mdp.id).value.should include @person2
       end
     end
    
