@@ -2,28 +2,35 @@ require 'spec_helper'
 
 describe MetaDatum do
 
-  context "a meta_key of type string" do
-
-    let :unsaved_meta_datum  do
-      md = MetaDatum.new 
-      md.meta_key = MetaKey.find_by_label("title")
-      md.value = "Blah"
-      md
-    end
-
-    it "should not raise an error upon save!" do
-      md = unsaved_meta_datum
-      expect {md.save!}.not_to raise_error
-    end
-
-    it "should be of type MetaDatumString after save!" do
-      pending "this will probalby never work; we need to change the way we create MetaDatum in the app"
-      md = unsaved_meta_datum
-      md.save!()
-      md.class.should == MetaDatumString
-    end
-
+  before :each do
+    @mk = MetaKey.create  meta_datum_object_type: "MetaDatumString"
   end
+
+  it "should raise an error when created without type and meta_key" do
+    expect {MetaDatum.reate }.to raise_error
+  end
+
+  it "should raise an error when instantiated without type and meta_key" do
+    expect {MetaDatum.reate }.to raise_error
+  end
+
+
+  it "should not raise an error when created with a type " do
+    expect {MetaDatum.create type: "MetaDatumString"}.not_to raise_error
+  end
+
+  it "should be of correct type when created with a type " do
+    (MetaDatum.create type: "MetaDatumString").class.should == MetaDatumString
+  end
+
+  it "should not raise an error when created with a meta_key" do
+    expect {MetaDatum.create meta_key: @mk}.not_to raise_error
+  end
+
+  it "should be of correct type when created with a meta_key" do
+    (MetaDatum.create meta_key: @mk).class.should == MetaDatumString
+  end
+
 
 end
 
